testbot: PowerOff() should not return an error if a VM is already off.
VMs may be put in unexpected states when an administrator manually
reconfigures them. So, even though we normally don't call PowerOff()
on powered off VMs, robustness dictates this change.  Also ensure that
if an error does occur the error message contains the VM name.  This
fixes various 'Scheduling problem in HandleXxx' error reports.

+++ testbot/lib/WineTestBot/VMs.pm
@@ -375,11 +375,21 @@ sub PowerOff($$)
   my ($ErrMessage, $Domain) = $self->_GetDomain();
   return $ErrMessage if (defined $ErrMessage);
 
-  eval { $Domain->destroy() };
-  return $@->message() if ($@);
-
-  return $self->UpdateStatus($Domain) if (!$NoStatus);
-  return undef;
+  if ($self->IsPoweredOn())
+  {
+    eval { $Domain->destroy() };
+    if ($@)
+    {
+      $ErrMessage = $@->message();
+    }
+    elsif ($self->IsPoweredOn())
+    {
+      $ErrMessage = "The VM is still active";
+    }
+  }
+  $ErrMessage ||= $self->UpdateStatus($Domain) if (!$NoStatus);
+  return undef if (!defined $ErrMessage);
+  return join("", "Could not power off ", $self->Name, ": ", $ErrMessage);
 }
 
 sub GetAgent($)
